+++ __test__/src/async_std.rs
@@ -1,6 +1,8 @@
 #[cfg(test)]
 mod tests {
 
+    use std::env::current_dir;
+
     use async_std::{fs::read_to_string, path::PathBuf};
 
     use get_dir::{
@@ -11,7 +13,7 @@ mod tests {
     async fn test_get_dir_by_target_dir() {
         let dir: PathBuf = GetDir::new()
             .targets(vec![Target::Dir(DirTarget { name: "src" })])
-            .get_async()
+            .run_async()
             .await
             .unwrap();
 
@@ -25,7 +27,7 @@ mod tests {
     async fn test_get_dir_by_target_file() {
         let dir: PathBuf = GetDir::new()
             .targets(vec![Target::File(FileTarget { name: "Cargo.toml" })])
-            .get_async()
+            .run_async()
             .await
             .unwrap();
 
@@ -36,10 +38,10 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_get_dir_by_target_reverse_dir() {
+    async fn test_get_dir_by_tarrun_reverse_dir() {
         let dir: PathBuf = GetDir::new()
             .targets(vec![Target::Dir(DirTarget { name: "target" })])
-            .get_reverse_async()
+            .run_reverse_async()
             .await
             .unwrap();
 
@@ -50,10 +52,10 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_get_dir_by_target_reverse_file() {
+    async fn test_get_dir_by_tarrun_reverse_file() {
         let dir: PathBuf = GetDir::new()
             .targets(vec![Target::File(FileTarget { name: "LICENSE" })])
-            .get_reverse_async()
+            .run_reverse_async()
             .await
             .unwrap();
 
@@ -61,5 +63,19 @@ mod tests {
             read_to_string(dir.join("Cargo.toml")).await.unwrap();
 
         assert!(content.contains("[workspace.dependencies]"));
+    }
+
+    #[tokio::test]
+    async fn test_get_dir_by_target_file_in_specific_dir() {
+        let dir: PathBuf = GetDir::new()
+            .directory(current_dir().unwrap().join("..").join("package"))
+            .targets(vec![Target::File(FileTarget { name: "lib.rs" })])
+            .run_async()
+            .await
+            .unwrap();
+
+        let content: String = read_to_string(dir.join("lib.rs")).await.unwrap();
+
+        assert!(content.contains("# Get Dir"));
     }
 }