+++ __test__/src/tokio.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod tests {
 
-    use std::path::PathBuf;
+    use std::{env::current_dir, path::PathBuf};
 
     use get_dir::{
         tokio::GetDirAsyncExt, DirTarget, FileTarget, GetDir, Target,
@@ -12,7 +12,7 @@ mod tests {
     async fn test_get_dir_by_target_dir() {
         let dir: PathBuf = GetDir::new()
             .targets(vec![Target::Dir(DirTarget { name: "src" })])
-            .get_async()
+            .run_async()
             .await
             .unwrap();
 
@@ -26,7 +26,7 @@ mod tests {
     async fn test_get_dir_by_target_file() {
         let dir: PathBuf = GetDir::new()
             .targets(vec![Target::File(FileTarget { name: "Cargo.toml" })])
-            .get_async()
+            .run_async()
             .await
             .unwrap();
 
@@ -37,10 +37,10 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_get_dir_by_target_reverse_dir() {
+    async fn test_get_dir_by_tarrun_reverse_dir() {
         let dir: PathBuf = GetDir::new()
             .targets(vec![Target::Dir(DirTarget { name: "target" })])
-            .get_reverse_async()
+            .run_reverse_async()
             .await
             .unwrap();
 
@@ -51,10 +51,10 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_get_dir_by_target_reverse_file() {
+    async fn test_get_dir_by_tarrun_reverse_file() {
         let dir: PathBuf = GetDir::new()
             .targets(vec![Target::File(FileTarget { name: "LICENSE" })])
-            .get_reverse_async()
+            .run_reverse_async()
             .await
             .unwrap();
 
@@ -62,5 +62,19 @@ mod tests {
             read_to_string(dir.join("Cargo.toml")).await.unwrap();
 
         assert!(content.contains("[workspace.dependencies]"));
+    }
+
+    #[tokio::test]
+    async fn test_get_dir_by_target_file_in_specific_dir() {
+        let dir: PathBuf = GetDir::new()
+            .directory(current_dir().unwrap().join("..").join("package"))
+            .targets(vec![Target::File(FileTarget { name: "lib.rs" })])
+            .run_async()
+            .await
+            .unwrap();
+
+        let content: String = read_to_string(dir.join("lib.rs")).await.unwrap();
+
+        assert!(content.contains("# Get Dir"));
     }
 }